# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Next Generation of Auth microservice integrated with kong for registration and issuing tokens",
        "title": "Open Tree Hole Auth",
        "contact": {
            "name": "Maintainer Chen Ke",
            "url": "https://danxi.fduhole.com/about",
            "email": "dev@fduhole.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/shamir": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shamir"
                ],
                "summary": "list related shamir PGP messages, admin only",
                "parameters": [
                    {
                        "type": "string",
                        "description": "recipient uid",
                        "name": "identity_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apis.PGPMessageResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    }
                }
            }
        },
        "/shamir/key": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shamir"
                ],
                "summary": "upload all PGP PublicKeys for encryption, admin only",
                "parameters": [
                    {
                        "description": "public keys",
                        "name": "public_keys",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.UploadPublicKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/utils.MessageResponse"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "data": {
                                                "$ref": "#/definitions/apis.IdentityNameResponse"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    }
                }
            }
        },
        "/shamir/shares": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shamir"
                ],
                "summary": "upload all shares of all users, cached",
                "parameters": [
                    {
                        "description": "shares",
                        "name": "shares",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.UploadSharesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.MessageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.IdentityNameResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.MessageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.IdentityNameResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    }
                }
            }
        },
        "/shamir/status": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shamir"
                ],
                "summary": "get shamir info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apis.ShamirStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    }
                }
            }
        },
        "/shamir/update": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shamir"
                ],
                "summary": "trigger for updating shamir",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    }
                }
            }
        },
        "/shamir/{user_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shamir"
                ],
                "summary": "get shamir PGP message, admin only",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target UserID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "identity_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apis.PGPMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "非管理员",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apis.IdentityNameResponse": {
            "type": "object",
            "properties": {
                "identity_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "apis.PGPMessageResponse": {
            "type": "object",
            "properties": {
                "pgp_message": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "apis.ShamirStatusResponse": {
            "type": "object",
            "properties": {
                "fail_message": {
                    "type": "string"
                },
                "new_public_keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ShamirPublicKey"
                    }
                },
                "now_user_id": {
                    "type": "integer"
                },
                "shamir_update_ready": {
                    "type": "boolean"
                },
                "shamir_updating": {
                    "type": "boolean"
                },
                "uploaded_shares_identity_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "apis.UploadPublicKeyRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "all standalone public keys",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "apis.UploadSharesRequest": {
            "type": "object"
        },
        "models.ShamirPublicKey": {
            "type": "object",
            "properties": {
                "identity_name": {
                    "type": "string"
                }
            }
        },
...
        "utils.MessageResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}